Support position assertions
Position assertions are syntax sugar for matching position of a value on
TOS. This can be used for conveniently snipping out individual values
out of arrays, or in expressions such as 'entry ?10'. Ticket #13
mentions some related stuff, and in particular an inconvenience with
accessing location expression values, where position asserions are going
to be useful.

Eventually this could be used for accessing script arguments passed on
command line.

Signed-off-by: Petr Machata <pmachata@gmail.com>