dwgrep: Fix snipping of first sentence in --help
When dwgrep shows --help, it snips out the first sentence of option's
docstring, and uses it as a short description of the command line
argument. The current logic assumes that sentences end in the first
period. However, "--version" uses a period for a string "MAJOR.MINOR",
where the period doesn't end the sentence.

So modigy the logic to ignore periods that are not followed by a white
space.

Signed-off-by: Petr Machata <pmachata@gmail.com>