lexer: Don't allow int literal next to a word
The lexer knows too much about how it is reasonable to construct an
integer literal, and stops scanning a token as soon as the digits that
make it up stop matching the numerical prefix. This e.g. 0b012 would be
parsed as two tokens: a binary integer literal 0b01, followed by a
decimal integer literal 2. Clearly that's wrong.

Thus change the lexer to be very optimistic about what an integer
literal might be. Basically anything that starts with a decimal digit is
one. (Note that hexadecimal numbers have a prefix of 0x, and thus start
with a decimal digit as well.) Leave it for further stages to decide
whether a given token actually is a valid number.

Signed-off-by: Petr Machata <pmachata@gmail.com>