parser: Improve message for overlarge int literal
std::out_of_range is thrown for overlarge literals, and the message is
just the name of the function where the overflow happened. When such an
exception bubbles up to the user, the message is not very helpful. So
catch the exception early and convert it to a more general exception
type with a better description.

Signed-off-by: Petr Machata <pmachata@gmail.com>