Privatize tree::build_pred
tree::build_exec was made private to enable passing arguments that are
not of interest to outside callers. For similars reasons it's desirable
to make build_pred private. Additionally, tree::build_pred is not called
besides from build_exec anyway, so it can be dropped from
tree::build_pred altogether, and completely privatized in an anonymous
namespace.

Signed-off-by: Petr Machata <pmachata@gmail.com>