WIP: Drop scope.hh and related stuff
This may reemerge from the current "struct bindings" in build.cc, but
currently there's no need for scopes in this form, so drop it and clean
up all uses. link_scopes goes away altogether. promote_scopes was
rewritten, but I suspect it can go as well (only some parser tests that
specifically verify that this is done fail, but since we may not care
this is done, ...)