WIP: Initial stab at closure compilation
This is a complete-ish (though buggy) implementation of the plan
described in the large comment in build.cc. All elements are in: the
rendezvous, reworked op_lex_closure, reworked op_apply, reworked
value_closure, pseudo bindings. The motivation example actually works
and is included as a test case.

However it's buggy and actually fails in more cases than before (9
instead of 7).