WIP: Convert op_subx, new op_inner_origin
op_inner_origin is like op_origin, but is part of the upstream chain for
purposes of state allocation only. Inner ops can't use nested states,
because a lot of the logic around origins relies on the fact that state
buffer is contiguously allocated, and that it's possible to compute
state of another operation on the outside of the nested state. Inner
origins help with that: all ops now really are in one chain.

Added some more tests, but a couple of them currently fail, because
subexpression assertions do allocate separate stacks and that doesn't
work.