WIP: Introduce layout
The way state allocation is done currently is heavy in pointer
arithmetic, and relies on the fact that it's possible to compute end of
state space from any state. That precludes having "side branches": a
side branch has its own origin, and therefore thinks that state space
end is there, but that's just the end of its own state subspace. That's
the reason op_inner_origin was introduced in that previous commit.

One could go ahead like this and convert gradually the whole codebase so
that all operations that share state space form one long chain. But
really there's a better way: pass around base pointer, and assign to
each operation a slot in the state space.

That's what the "layout" and (tentatively named) scon2 do.
Operations (and predicated) that have state take "layout" as a ctor
argument, and ask the layout to assign them a slot, given state struct
type. Layout returns a location, which is stored in class data.

scon2 then maintains the state space. It's constructed form a layout,
and given a loc and a type, returns state struct. Since now every op
knows its absolute location in state space, a lot of code gets simpler.

The patch is huge mostly because it prompts changes across the codebase,
but the actual code is much simpler. Vast majority of pointer arithmetic
goes away.