Drop undocumented arrow binding syntax
There are plenty ways to bind variables, and this one isn't documented.
The advantage of this syntax was that it allowed recursive functions to
be easily expressed, but if that is desirable, let's have a "let rec"
alongside "let" or similar construct, and not something this random.

Convert the three existing test cases to the let form (where the
recursion is a bit more awkward to express, but still supported).

Signed-off-by: Petr Machata <pmachata@gmail.com>