Introduce bindings & layout
Two major changes are introduced to address issue #25: individual
operations become constant, and any state that they need for operation
is carried in a separate state container; and the correspondence between
read and bind operations is now determined at the time of query
compilation, and each read knows its bind right away.

In order to track which part of the state space is used by which op,
introduce class layout. An instance of that class is passed to ops that
have state at compile time, and these ops ask the layout to reserve
space for their state. Layout then hands out a token that represents
location is a state buffer, which the op then uses to get at its state,
given a state buffer.

In order to track the correspondence between variable names and binding
ops, introduce class bindings and class uprefs (the latter is for
handling references to scopes out of current block).

Modify build.cc to pass the current layout object, current binding map,
and current upvalue map, among the various builder functions. Update the
places that call into the builder to pass in the layout as well.

Specifically to support upvalues, recognize that there needs to be a
special reserved area at each state buffer, where reference to upvalue
vector is stored. This area is called rendezvous, and the token to
access it is passed in rdv_ll.

Signed-off-by: Petr Machata <pmachata@mellanox.com>