Introduce per-op state
To fix issue #25, read operations will newly directly know which bind
operation binds the value that they are reading, and there won't be a
frame attached to each stack. That's a problem for closures, which can
reference surrouding context that needs to be closed over, and where
frames worked nicely.

Thus, because closures need to be revamped anyway, this patch also gets
rid of removing the explicit calls to tree::build_exec inside op_apply.
Instead, build_exec will compile the contents of a closure ahead of
time, and pass the op chain to op_lex_closure, which will pass it on to
value_closure, where op_apply will pick it up.

But this causes problems for recursive closures, where the same chain is
running in several activations. State thus can't be kept at ops, but
needs to be moved out of them.

Therefore continue the recent constification work by making op::next
const. To make this possible, extract state of current ops to a separate
class, and introduce two new calls: op::state_con and op::state_des,
that serve the role of constructor and destructor of state of a given
activation. Introduce "scon", state container, that holds state buffer
of an op chain. And introduce scon_guard to manage construction and
destruction of state in that state buffer.

This is currently not yet leveraged, op_apply still uses the old
approach, it just converts the codebase to make the follow-up changes
easier.

Signed-off-by: Petr Machata <pmachata@gmail.com>