Rework reads and closures
In this second part of fix to issue #25, reads now directly access the
corresponding binds to ask them what the variable value is. This prompts
updates to handling of closures, because reads of their environment need
to be translated differently, and the environment needs to be captured
and stored at the closure.

This does allow some nice cleanups: scopes become free, so some of the
scope-linking and scope-promotion code goes away, frames are not needed
anymore, etc. Unbound reads are now noticed eagerly instead of in
runtime.

Signed-off-by: Petr Machata <pmachata@gmail.com>