Make scon_guard movable
scon_guard needs to have a forbidden copy constructor, otherwise
temporary copies end up calling state_des before we are ready to get rid
of that state. In order to be usable in nonstd::optional, it then needs
to have a move constructor.

So convert m_op to a pointer so that there's a way to mark that a
scon_guard was gutted and is not valid any more. Mark copy constructor
as deleted, and implement a move constructor.

Signed-off-by: Petr Machata <pmachata@gmail.com>