Share state for cond, then and else of op_ifelse
op_ifelse needs to reinitialize the state space of the condition op
before it starts pulling stacks from there, because it never drains the
op chain (which would bring it to init state), but gives up after
getting the first stack.

That means that the cond op state space is free for reuse by whichever
branch is chosen for the body. And of course only one of next and else
will be instantiated then.

So the three op chains in flight are never active at the same time.
Therefore, initialize the state space as union of the three.

Additionally, change the ifelse state to carry the responsibility for
destroying the state space of the selected body op chain, i.e. don't run
con and des by hand, but defer to a scon_guard (which is optional,
because it only gets initialized after cond is run).

Signed-off-by: Petr Machata <pmachata@gmail.com>