parser: Parse infix cmp as let/let/read
In order to remove references to builtins from the parser and shift them
all to builder, something needs to be done about the way the operators
are parsed. Since scopes are free and bindings almost free, there's
little reason to hold support for specifically this syntax. Express
the form:
  (A <op> B)
as
  (scope (let ~a~ := A; let ~b~ := B; (read ~a~) (read ~b~) (read <op>)))

Since the expanded form is so complex, it's probably not useful to test
it in test-parser. Just defer to the runtime tests to determine it's
doing the right thing.

Signed-off-by: Petr Machata <pmachata@gmail.com>