doc: syntax: Change infix operator explanation
The previous documentation makes a strong connection between an infix
operator and a regular word with the same behavior. That woud seem to
imply that overriding, say, ?eq would have impact on behavior of ==.

But such an association is hard to do in extensible manner. Imagine that
there is an "alias" sort of entry in a vocabulary. Whether two words are
aliased can't depend on stack profile, because doing lookups mid-runtime
is not how Zwerg is meant to operatre. Thus the association needs to be
fixed and static. But that means that if two vocabularies introduce the
same operator, the two aliases will likely be incompatible with one
another, and it will be impossible to reconcile the difference. That's
unlike operators themselves. Those are just words, and can be overloaded
at will, and the individual overloads likely depend on types introduced
by the vocabulary in question. Thus it's likely that the overloads do
not clash and it's possible to merge the individual definitions.

Thus change the explanation to make it clear that OP2 is mentioned just
for explanation and there's no actual connection between the two, except
that they behave the same way. This change can't impact programs out
there, because so far it wasn't even possible to override builtins.

Signed-off-by: Petr Machata <pmachata@gmail.com>