parser: Allow let to take literal names
This is meant for experimentation with whether it makes sense to allow
overloading of operators. Hence the weird syntax (for a real feature, we
would accept TOK_OP rather then TOK_LIT_STR, and IdListOpt would be
extended to allow mix of words and ops, so that other binding constructs
take this as well).

Signed-off-by: Petr Machata <pmachata@gmail.com>