libzwerg: Convert builtin invocations to reads
This patch changes the way builtins are looked up: instread of deciding,
in the lexer, that a particular word is a builtin reference, all word
references are just reads. Later on the builder decides whether a
particular read references a binding or a builtin.

To facilitate this change, bindings and uprefs now come in two sorts:
builtin and binding. All builtins are then imported into toplevel scope.
In follow-up patches, this will allow overriding of builtin words in
more inner scopes.

Signed-off-by: Petr Machata <pmachata@gmail.com>