libzwerg: builtin-dw-voc: known-dwarf.h misses DW_MACRO_GNU_*
known-dwarf.h, both as we build it, and as elfutils ships it, misses the
definitions of DW_MACRO_GNU_* enumerators, because as of elfutils 0.170
they are not enums anymore. Since the old enum is now kept in
backward-compatibility mode, it's likely it won't be added to, and it's
safe to expand it in situ in dwgrep.

Signed-off-by: Petr Machata <pmachata@gmail.com>