libzwerg/op: Capture needs to reset its subop
When capture is done draining its subop for the current list instance,
it needs to des/con its subop's state. If the subop is an alt list,
after the draining it knows that there's nothing more to yield so it
won't. Capture needs to reset it, thereby signalling that it's safe to
pull from upstream again.

Signed-off-by: Petr Machata <pmachata@gmail.com>