libzwerg/atval: Don't crash on unhandled artifacts
Not being able to decode a value is a problem, and in general the
computation may not be able to proceed. However crashing is not
acceptable. Instead, raise an exception with a suitable message, which
communicates the error message to the caller application.

Add the actual value of unhandled artifact to the message. Since that
part is handled by known-dwarf, the name may ba available even if the
attribute value decoder wasn't updated to handle that artifact.

Signed-off-by: Petr Machata <pmachata@gmail.com>