libzwerg/test-dw: Factor out test_pairs
Running a little script that yields a number of value/key pairs looks
like a useful way to test all sorts of features. Convert test_defaulted
to a helper function that services tests of this type. Convert
test_defaulted to use constants instead of mere unsigneds, and dispatch
to the new helper.

Signed-off-by: Petr Machata <pmachata@gmail.com>