libzwerg: Split ELF parts from vocabulary_dw
The ELF vocabulary is shaping up to large enough that it's reasonable to
keep it in a vocabulary of its own. The ELF symbol stuff that's been
around for a bit belongs there as well. Introduce ELF-related API
header, move tests from test-dw to a new file test-elf, introduce a
vocabulary builder, and have dwgrep.cc import the ELF vocabulary by
default so that the queries keep working.