libzwerg: elf: Change @foo and @elffoo to efoo
In Dwarf, the @ notation is used for attributes that can be attached to
T_ENTRY objects. There can be a number of them, and it's not
predetermined which ones those will be. Correspondingly there are also
the ? and ! forms for querying whether a given attribute is available.

The properties of T_ELF objects are unlike that. Therefore don't use @
to access them. Instead, use "e" as a prefix to disambiguate ("type" is
not an overload), and so drop the "elf"-prefixed variants as well.

Signed-off-by: Petr Machata <pmachata@gmail.com>