libzwerg: Introduce builtin-foo-voc.hh
So far, dwgrep_vocabulary_foo() that is defined in builtin-foo-voc.cc,
was declared in builtin-foo.hh. That's unexpected and confusing, and
unnecessary, and the declaration is very frugal and doesn't need any
includes.

Therefore add a header file purely for the vocabulary-generating
function, and move the declaration from the main .hh file there.

Signed-off-by: Petr Machata <pmachata@gmail.com>