libzwerg/elfcst: Generalize show_unknown
show_unknown currently hard-codes two ranges of "unknown" values: OS and
PROC. DT has two more ranges that might be formatted this way.

Generalize the unknown printer so that it can be called with one range,
formats it if the code falls within the range, and returns true indicating
that the value was shown. Chain multiple such calls, falling back to a
last-resort unknown formatter.

Signed-off-by: Petr Machata <me@pmachata.org>