libzwerg/selector: Rewrite assertion on value_type::code() return type
The assertion is in place because the selector relies on the fact that
it can lay out the stact fingerprint in a certain way. However on a newer
GCC, it started warning:

 libzwerg/selector.hh:45:57: warning: ‘this’ pointer is null [-Wnonnull]
    45 |   static_assert (sizeof (((value_type *) nullptr)->code ()) == 1,

While it does not break the build, and is a spurious warning, it is noisy
and annoying. Rewrite the assertions to be inside an unused private method
that takes a bogus value_type argument, and use that to do the query.

Signed-off-by: Petr Machata <me@pmachata.org>