libzwerg/op: op_subx::next: Rewrite one while loop
Current GCC warns about the naked if (st.m_stk = ...). Convert into a more
canonical form where the assignment is done separately from the if. Then
notice that the second iteration of the while loop is guaranteed to
terminate: either st.m_stk is null after the assignment, and we bail out.
Or it's non-null, and the while stops. Therefore convert to an if.

Signed-off-by: Petr Machata <me@pmachata.org>